Review remarks
Changed the code a bit so that the Validator Management tool is the 'single source of truth' and is the sole generator of the error messages.
The ValidatorErrorsToMarkdown will generate the description, properties, etc and replace a section in the file that is generated by the management tool. This way the order is fixed and it's clear for users where auto-generated stuff is.