Transfer Details & Example code from errormessages to markdown files (#143)
* removed details and examples from auto-genrated content

* Update TestGenerator.cs

* Update TestGenerator.cs

* Update TestGenerator.cs

* Update TestGenerator.cs

* Added base documentation for existing checks

* Update ParseXmlToMarkDown.cs

* Removed examples and details from source

* Review remarks

Changed the code a bit so that the Validator Management tool is the 'single source of truth' and is the sole generator of the error messages.
The ValidatorErrorsToMarkdown will generate the description, properties, etc and replace a section in the file that is generated by the management tool. This way the order is fixed and it's clear for users where auto-generated stuff is.

* Small tweaks

* Add interface properties back, but mark them obsolete. This way no syntax breaking changes are introduced

---------

Co-authored-by: Michiel Oda <michiel.oda@skyline.be>