Update uom.xsd
I could see that in EPM system, there are a lot of columns for which devices or events type was used, and validator is suppressed, so I think having these two units available would bring some value. Also, not sure if this is the only place units should be added, but for the easy list maintenance I want to believe that it is.